refactor: better feature scoping, fix turso compilation, prune dependency tree for library consumers
