+++ cargo-shuttle/Cargo.toml
@@ -9,9 +9,9 @@ homepage = "https://www.shuttle.rs"
 rust-version = "1.70"
 
 [dependencies]
-shuttle-common = { workspace = true, features = ["models"] }
-shuttle-proto = { workspace = true }
-shuttle-service = { workspace = true, features = ["builder"] }
+shuttle-common = { workspace = true, features = ["models", "backend"] }
+shuttle-proto = { workspace = true, features = ["provisioner", "runtime"] }
+shuttle-service = { workspace = true, features = ["builder", "runner"] }
 
 anyhow = { workspace = true }
 async-trait = { workspace = true }
