+++ deployer/src/runtime_manager.rs
@@ -13,9 +13,9 @@ use shuttle_common::{
 };
 use shuttle_proto::{
     logger::{logger_client::LoggerClient, Batcher, LogItem, LogLine},
-    runtime::{self, runtime_client::RuntimeClient, StopRequest},
+    runtime::{runtime_client::RuntimeClient, StopRequest},
 };
-use shuttle_service::Environment;
+use shuttle_service::{runner, Environment};
 use tokio::{io::AsyncBufReadExt, io::BufReader, process, sync::Mutex};
 use tonic::transport::Channel;
 use tracing::{debug, error, info, trace, warn};
@@ -125,7 +125,7 @@ impl RuntimeManager {
                 .join("bin/shuttle-next")
         };
 
-        let (mut process, runtime_client) = runtime::start(
+        let (mut process, runtime_client) = runner::start(
             is_next,
             Environment::Deployment,
             &self.provisioner_address,
