+++ gateway/src/project.rs
@@ -24,7 +24,8 @@ use once_cell::sync::Lazy;
 use rand::distributions::{Alphanumeric, DistString};
 use serde::{Deserialize, Serialize};
 use shuttle_common::backends::headers::{X_SHUTTLE_ACCOUNT_NAME, X_SHUTTLE_ADMIN_SECRET};
-use shuttle_common::models::project::{default_idle_minutes, ProjectName, DEFAULT_IDLE_MINUTES};
+use shuttle_common::constants::{default_idle_minutes, DEFAULT_IDLE_MINUTES};
+use shuttle_common::models::project::ProjectName;
 use shuttle_common::models::service;
 use tokio::time::{sleep, timeout};
 use tracing::{debug, error, info, instrument, trace, warn};
