+++ logger/Cargo.toml
@@ -7,7 +7,7 @@ repository.workspace = true
 
 [dependencies]
 shuttle-common = { workspace = true, features = ["backend", "tonic"] }
-shuttle-proto = { workspace = true }
+shuttle-proto = { workspace = true, features = ["logger"] }
 
 async-trait = { workspace = true }
 chrono = { workspace = true }
