+++ proto/Cargo.toml
@@ -6,23 +6,29 @@ license.workspace = true
 description = "Library for all the gRPC definitions used by shuttle"
 
 [dependencies]
-shuttle-common = { workspace = true, features = [
-    "claims",
-    "service",
-    "wasm",
-    "models",
-    "backend",
-] }
+shuttle-common = { workspace = true }
 
-anyhow = { workspace = true }
-chrono = { workspace = true }
-dunce = { workspace = true }
+anyhow = { workspace = true, optional = true }
+chrono = { workspace = true, optional = true }
 futures-core = "0.3.28"
-home = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
-tokio = { workspace = true, features = ["process"] }
+tokio = { workspace = true, optional = true }
 tonic = { workspace = true }
-tower = { workspace = true }
-tracing = { workspace = true }
-serde_json = { workspace = true }
+tracing = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+
+[features]
+default = []
+
+builder = []
+logger = [
+    "shuttle-common/service",
+    "chrono",
+    "tracing",
+    "tokio/macros",
+    "tokio/time",
+]
+provisioner = []
+resource-recorder = ["anyhow", "serde_json"]
+runtime = []
