+++ proto/src/lib.rs
@@ -1,5 +1,11 @@
 mod generated;
 
+// useful re-exports if types are needed in other crates
+pub use prost;
+pub use prost_types;
+pub use tonic;
+
+#[cfg(feature = "provisioner")]
 pub mod provisioner {
     use std::fmt::Display;
 
@@ -92,106 +98,12 @@ pub mod provisioner {
     }
 }
 
+#[cfg(feature = "runtime")]
 pub mod runtime {
-    use std::{
-        path::{Path, PathBuf},
-        process::Stdio,
-        time::Duration,
-    };
-
-    use anyhow::Context;
-    use shuttle_common::{
-        claims::{ClaimLayer, ClaimService, InjectPropagation, InjectPropagationLayer},
-        deployment::Environment,
-    };
-    use tokio::process;
-    use tonic::transport::{Channel, Endpoint};
-    use tower::ServiceBuilder;
-    use tracing::{info, trace};
-
     pub use super::generated::runtime::*;
-
-    pub async fn start(
-        wasm: bool,
-        environment: Environment,
-        provisioner_address: &str,
-        auth_uri: Option<&String>,
-        port: u16,
-        runtime_executable: PathBuf,
-        project_path: &Path,
-    ) -> anyhow::Result<(
-        process::Child,
-        runtime_client::RuntimeClient<ClaimService<InjectPropagation<Channel>>>,
-    )> {
-        let port = &port.to_string();
-        let environment = &environment.to_string();
-
-        let args = if wasm {
-            vec!["--port", port]
-        } else {
-            let mut args = vec![
-                "--port",
-                port,
-                "--provisioner-address",
-                provisioner_address,
-                "--env",
-                environment,
-            ];
-
-            if let Some(auth_uri) = auth_uri {
-                args.append(&mut vec!["--auth-uri", auth_uri]);
-            }
-
-            args
-        };
-
-        info!(
-            "Spawning runtime process: {} {}",
-            runtime_executable.display(),
-            args.join(" ")
-        );
-        let runtime = process::Command::new(
-            dunce::canonicalize(runtime_executable).context("canonicalize path of executable")?,
-        )
-        .current_dir(project_path)
-        .args(&args)
-        .stdout(Stdio::piped())
-        .kill_on_drop(true)
-        .spawn()
-        .context("spawning runtime process")?;
-
-        info!("connecting runtime client");
-        let conn = Endpoint::new(format!("http://127.0.0.1:{port}"))
-            .context("creating runtime client endpoint")?
-            .connect_timeout(Duration::from_secs(5));
-
-        // Wait for the spawned process to open the control port.
-        // Connecting instantly does not give it enough time.
-        let channel = tokio::time::timeout(Duration::from_millis(7000), async move {
-            let mut ms = 5;
-            loop {
-                if let Ok(channel) = conn.connect().await {
-                    break channel;
-                }
-                trace!("waiting for runtime control port to open");
-                // exponential backoff
-                tokio::time::sleep(Duration::from_millis(ms)).await;
-                ms *= 2;
-            }
-        })
-        .await
-        .context("runtime control port did not open in time")?;
-
-        let channel = ServiceBuilder::new()
-            .layer(ClaimLayer)
-            .layer(InjectPropagationLayer)
-            .service(channel);
-        let runtime_client = runtime_client::RuntimeClient::new(channel);
-
-        Ok((runtime, runtime_client))
-    }
 }
 
+#[cfg(feature = "resource-recorder")]
 pub mod resource_recorder {
     use anyhow::Context;
     use std::str::FromStr;
@@ -238,10 +150,12 @@ pub mod resource_recorder {
     }
 }
 
+#[cfg(feature = "builder")]
 pub mod builder {
     pub use super::generated::builder::*;
 }
 
+#[cfg(feature = "logger")]
 pub mod logger {
     use std::str::FromStr;
     use std::time::Duration;
@@ -260,8 +174,6 @@ pub mod logger {
         log::{Backend, LogItem as LogItemCommon, LogRecorder},
         DeploymentId,
     };
-
-    use self::logger_client::LoggerClient;
 
     pub use super::generated::logger::*;
 
@@ -333,7 +245,7 @@ pub mod logger {
     }
 
     #[async_trait]
-    impl<T> VecReceiver for LoggerClient<T>
+    impl<T> VecReceiver for logger_client::LoggerClient<T>
     where
         T: tonic::client::GrpcService<tonic::body::BoxBody> + Send + Sync + Clone,
         T::Error: Into<StdError>,
