+++ service/Cargo.toml
@@ -12,19 +12,24 @@ doctest = false
 
 [dependencies]
 shuttle-common = { workspace = true, features = [
+    "claims",
     "service",
     "tracing",
     "display",
 ] }
+shuttle-proto = { workspace = true, optional = true, features = ["runtime"] }
 
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 cargo_metadata = { workspace = true, optional = true }
+dunce = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
 strfmt = { workspace = true }
 thiserror = { workspace = true }
-tokio = { workspace = true, features = ["process", "io-util", "sync", "time"], optional = true }
-toml = { workspace = true, optional = true  }
+tokio = { workspace = true, optional = true }
+tonic = { workspace = true, optional = true }
+tower = { workspace = true, optional = true }
+toml = { workspace = true, optional = true }
 tracing = { workspace = true, optional = true }
 
 [dev-dependencies]
@@ -33,4 +38,13 @@ tokio = { workspace = true, features = ["macros", "rt"] }
 [features]
 default = []
 
-builder = ["cargo_metadata", "tokio", "toml", "tracing"]
+builder = [
+    "cargo_metadata",
+    "tokio/process",
+    "tokio/io-util",
+    "tokio/sync",
+    "tokio/time",
+    "toml",
+    "tracing",
+]
+runner = ["shuttle-proto", "tokio/process", "dunce", "tonic", "tower"]
